feat(sensor): #80 (2)
merge start/stop commands into one file
