+++ sensor/cli/commands/__init__.py
@@ -1,5 +1,3 @@
 from .enclosure import enclosure
 from .info import info
-from .is_running import is_running
-from .start import start
-from .stop import stop
+from .background_process import start, is_running, stop
