+++ sensor/cli/main.py
@@ -16,8 +16,8 @@ def cli() -> None:
 
 cli.add_command(commands.enclosure, name="enclosure")
 cli.add_command(commands.info, name="info")
-cli.add_command(commands.is_running, name="is-running")
 cli.add_command(commands.start, name="start")
+cli.add_command(commands.is_running, name="is-running")
 cli.add_command(commands.stop, name="stop")
 
 if __name__ == "__main__":
